requirements.txt
autoenv==1.0.0
binaryornot==0.4.0
chardet==2.3.0
click==6.3
cloudinary==1.4.0
cookiecutter==1.3.0
coreapi==1.32.0
diff-match-patch==20121119
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.4
django-cors-headers==1.1.0
django-extensions==1.7.1
django-filter==0.13.0
django-import-export==0.4.5
django-rest-swagger==0.3.8
django-toolbelt==0.0.1
djangocms-installer==0.8.7
djangorestframework==3.4.1
djangorestframework-csv==1.4.1
drf-extensions==0.2.8
future==0.15.2
gunicorn==19.6.0
itypes==1.1.0
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
models==0.9.3
openapi-codec==1.0.0
Pillow==3.3.0
psycopg2==2.6.2
pydot==1.2.2
pydotplus==2.0.2
pyparsing==2.1.8
pytz==2016.3
PyYAML==3.11
qrcode==5.1
requests==2.10.0
rest-framework-generic-relations==1.1.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.11.7
simplejson==3.8.2
six==1.10.0
static3==0.7.0
tablib==0.11.2
tzlocal==1.2.2
uritemplate==0.6
whichcraft==0.1.1
autoenv==1.0.0
binaryornot==0.4.0
chardet==2.3.0
click==6.3
cloudinary==1.4.0
cookiecutter==1.3.0
coreapi==1.32.0
diff-match-patch==20121119
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.4
django-cors-headers==1.1.0
django-extensions==1.7.1
django-filter==0.13.0
django-import-export==0.4.5
django-rest-swagger==0.3.8
django-toolbelt==0.0.1
djangocms-installer==0.8.7
djangorestframework==3.4.1
djangorestframework-csv==1.4.1
drf-extensions==0.2.8
future==0.15.2
gunicorn==19.6.0
itypes==1.1.0
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
models==0.9.3
openapi-codec==1.0.0
Pillow==3.3.0
psycopg2==2.6.2
pydot==1.2.2
pydotplus==2.0.2
pyparsing==2.1.8
pytz==2016.3
PyYAML==3.11
qrcode==5.1
requests==2.10.0
rest-framework-generic-relations==1.1.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.11.7
simplejson==3.8.2
six==1.10.0
static3==0.7.0
tablib==0.11.2
tzlocal==1.2.2
uritemplate==0.6
whichcraft==0.1.1
